Implement Italian text post-processing for OCR
Add comprehensive post-processing pipeline to fix common OCR errors in
Italian text, matching Python implementation behavior exactly.

Features:
- Fix apostrophe character substitutions (ľ → ')
- Normalize typographic quotes and punctuation
- Separate compound words with missing apostrophes
- Fix contextual errors (missing accents, common mistakes)

Implementation:
- TextPostProcessor.cs: Complete post-processing pipeline with 4 stages
  1. NormalizePunctuation(): Convert typographic chars to ASCII
  2. FixApostrophes(): Fix ľ (U+013E) character substitutions
  3. SeparateCompoundWords(): Add missing apostrophes in compounds
  4. FixContextualErrors(): Correct common Italian OCR mistakes

- OcrEngine.cs: Integrated post-processing after recognition step
- 26 comprehensive test cases validate exact match with Python

Results:
- All apostrophe errors fixed (dell'albero, l'unico, etc.)
- All accent marks preserved (già, più, Perché, città)
- Compound words correctly separated (nell'università, dall'alba)
- Test accuracy: 26/26 tests pass (100%)

Cleanup:
- Removed debug logging from all components
- Removed temporary debug code and file outputs

+++ EasyOcrNet/Detection/CraftDetector.cs
@@ -211,11 +211,12 @@ public class CraftDetector : IDetector
         foreach (var box in boxes)
         {
             // box format: [[x0,y0], [x1,y1], [x2,y2], [x3,y3]]
+            // IMPORTANT: Keep float precision for accurate grouping!
             var boundingBox = new BoundingBox(
-                TopLeft: new Point2D((int)box[0][0], (int)box[0][1]),
-                TopRight: new Point2D((int)box[1][0], (int)box[1][1]),
-                BottomRight: new Point2D((int)box[2][0], (int)box[2][1]),
-                BottomLeft: new Point2D((int)box[3][0], (int)box[3][1])
+                TopLeft: new Point2D(box[0][0], box[0][1]),
+                TopRight: new Point2D(box[1][0], box[1][1]),
+                BottomRight: new Point2D(box[2][0], box[2][1]),
+                BottomLeft: new Point2D(box[3][0], box[3][1])
             );
 
             detections.Add(new DetectionResult(boundingBox, Confidence: 1.0f));
